The whole-body-complete instrument has ranked the last three rungs correctly, each
converting 1:1 with the bucket drop equalling the census gain exactly. It had one
blind spot, and D2 named it: `recv-object`'s 0.0% says only that the grammar does
not finish those bodies, not WHAT ELSE blocks them. Three forms hold 172,625
functions there, so the difference between "one shared blocker" and "three each"
was about two orders of magnitude in value and nothing measured could tell them
apart.

**The answer is no.** Only **15,962 of 172,625 (9.2%)** are one construct away;
21.0% of the bucket is reachable within four. Grant-count distribution: k=1 22,645,
k=2 25,588, k=3 4,528, k=4 2,152, >4 120,514, plus 51,810 whose pair and 39,474
whose form remain unmeasured. That redirects effort away from a target that looked
large and is not, which is what this rung existed to establish.

  form         total   dominant 2nd blocker              share   1 away  within 4
  recv-object  64,905  type-ptr 27,992 / branch-0x39 23,632  43/36%   2,418   11,084
  data-addr    56,634  plain-call 55,925                  98.8%   1,063   22,357
  recv-load    51,086  chain-bind 29,001                    57%  12,481   14,697

**It also inverts §14.7's ranking.** `data-addr x plain-call x type-ptr` is
**21,642** (20,579 at k=2 plus 1,063 at k=1) and is now the top pair — needing
REFHI/REFLO on data symbols, the `$SG` `.rdata` COMDAT, an ordinary call as a
value, and pointer-typed argument operands, with no new frame class and no basic
blocks. §14.7 had ranked data-symbol addressing LAST on the strength of that form's
0.0% single-form completeness: true of the form and misleading about the row.

The estimate is quoted at **15,000–20,000 against the 21,642 bound**, and §15.4's
"quote generated-code estimates at the bound" is explicitly NOT applied — that
correction was derived from compiler-generated destructors with no freedom in the
gated fields, and these are hand-written bodies.

**Unlooked-for, and the third instance of one pattern this session**: `chained`
undercounts chains by **~4.4x**. `mod.rs`'s statement dispatch eats a statement-head
`26 <tok>` as an assignment destination, so a chain in a *value* position files as
`recv-load`. True chain population **43,521 = 16.2% of the bucket**, against D2's
2.8%. Pinned by a probe pair differing only in the destination and producing
identical two-link IL. This is the mis-attribution failure one level up — inside the
classifier built to fix it the level below.

Acceptance unchanged, verified past the numerator: census exactly 280,020, **0
per-TU differences** in in-class count or class, all **924** non-`mcall` keys and
their sum identical, and the `expr-call-in-expr` family total identical at 268,140.
Sub-buckets sum to D2's parents exactly, with **19** functions re-decomposed out of
`op-0x99` — the one intentional walk change, accounted for individually.

Left UNMEASURED on purpose, and worth more than a guess: `op-0x9B` (**39,361**),
the largest item on the worklist whose size is unknown. Classifying it needs a
depth-0 `32` treated as non-decisive mid-statement, which would move functions on a
hypothesis. `op-0x80` (579) + `op-0x08` (234) are a labelled possible desync at 0.3%
of the bucket; the residue SHAPE argues against a systemic cause — 24 distinct
opcodes over 4.2% with 80% in one known byte, not the flat 197-bucket fingerprint
that exposed the LEB descriptor bug — but it is not ruled out. `branch-0x39`'s
polarity is undetermined; one probe pins `0x38` as branch-if-false.

One measure asymmetry, stated because it biases a number: pass 1 is D2's argument
grammar verbatim, so every `-whole` count reproduces exactly, while passes >= 2
admit granted constructs inside argument regions. `-whole<k>` is therefore measured
against a marginally looser frame than `-whole`.

110 fixtures, /Ox 45 /O1 42 /O2 42 /Ox /Gy 42, 0 mismatch in every lane, sweep
3,259/0, workload mismatch 0, census 280,020 unchanged by construction.

+++ crates/c2-il/src/func/body/shapes.rs
@@ -1750,7 +1750,7 @@ pub(crate) fn parse_call_shape(
                 ctx: "call-intrinsic",
                 byte: Some(0x40),
                 off: *p,
-                aux: id as u32,
+                aux: id as u64,
             });
         }
         return Err(blk(seg, *p, "call-token"));
